# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Next Home API
  description: Rental Homes Platform REST API
  version: "1.0.0"
  contact:
    name: Next Home Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.nexthome.app/api/v1
    description: Production (OCI)
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID token obtained from Firebase Auth SDK"
  schemas:
...
    Property:
      type: object
      properties:
        id:             { type: string, format: uuid }
        title:          { type: string }
        description:    { type: string }
        propertyType:   { type: string, enum: [apartment, house, villa, studio, commercial] }
        status:         { type: string, enum: [active, inactive, rented, maintenance] }
        address:        { type: string }
        city:           { type: string }
        state:          { type: string }
        latitude:       { type: number }
        longitude:      { type: number }
        rentPerMonth:   { type: number }
        securityDeposit: { type: number }
        bedrooms:       { type: integer }
        bathrooms:      { type: integer }
        areaSqft:       { type: number }
        furnishedStatus: { type: string, enum: [furnished, semi-furnished, unfurnished] }
        amenities:      { type: array, items: { type: string } }
        images:         { type: array, items: { type: string, format: uri } }
        landlordId:     { type: string, format: uuid }
...
paths:
...
  /properties:
    get:
      summary: List properties with filters
      tags: [Properties]
      parameters:
        - { name: city,          in: query, schema: { type: string } }
        - { name: propertyType,  in: query, schema: { type: string } }
        - { name: minRent,       in: query, schema: { type: number } }
        - { name: maxRent,       in: query, schema: { type: number } }
        - { name: bedrooms,      in: query, schema: { type: integer } }
        - { name: lat,           in: query, schema: { type: number } }
        - { name: lng,           in: query, schema: { type: number } }
        - { name: radiusKm,      in: query, schema: { type: number } }
        - { name: page,          in: query, schema: { type: integer, default: 1 } }
        - { name: limit,         in: query, schema: { type: integer, default: 20 } }
      responses:
        "200": { description: Property list }
    post:
      summary: Create property (landlord only)
      tags: [Properties]
      security: [{ FirebaseAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Property" }
      responses:
        "201": { description: Property created }
  /properties/{id}:
    get:
      summary: Get property detail
      tags: [Properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Property detail }
...